FlashCard finished

+++ src/components/AdminDashboard/Content&Resources/content/bank/MCQBank/McqBankCardForAdmin.tsx
@@ -1,7 +1,10 @@
 import CommonButton from "@/common/button/CommonButton";
 import { toBerhanTime } from "@/help/help";
+import { useDeleteFlashCardBankMutation } from "@/store/features/adminDashboard/ContentResources/flashCard/flashCardSlice";
 import { useDeleteMcqBankApiMutation } from "@/store/features/adminDashboard/ContentResources/MCQ/mcqApi";
 import { SingleMcqBank } from "@/store/features/adminDashboard/ContentResources/MCQ/type/allContent";
+import { useAppSelector } from "@/store/hook";
+import { RootState } from "@/store/store";
 import { FC } from "react";
 
 interface Props {
@@ -9,11 +12,30 @@ interface Props {
   setMcqBankId: (id: string) => void;
 }
 const McqBankCardForAdmin: FC<Props> = ({ data, setMcqBankId }) => {
+  const { contentType } = useAppSelector(
+    (state: RootState) => state.staticContent
+  );
   const [deleteMcqBankApi, { isLoading }] = useDeleteMcqBankApiMutation();
+  const [deleteFlashCardBank, { isLoading: isDeleting }] =
+    useDeleteFlashCardBankMutation();
+
+  // useDeleteFlashCardBankMutation,
+  // useDeleteSingleFlashCardMutation,
 
   const handleDelete = async (id: string) => {
-    await deleteMcqBankApi(id);
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this MCQ bank?"
+    );
+    if (!confirmDelete) return;
+    if (contentType === "MCQ") {
+      await deleteMcqBankApi(id);
+    }
+
+    if (contentType === "Flashcard") {
+      await deleteFlashCardBank(id);
+    }
   };
+
   return (
     <div className=" bg-white rounded-3xl  border border-border overflow-hidden hover:shadow-xl transition-all duration-300">
       {/* Header */}
@@ -68,7 +90,7 @@ const McqBankCardForAdmin: FC<Props> = ({ data, setMcqBankId }) => {
           onClick={() => handleDelete(data._id)}
           className=" bg-red-500 !text-white"
         >
-          {isLoading ? "Deleting..." : "Delete"}
+          {isLoading || isDeleting ? "Deleting..." : "Delete"}
         </CommonButton>
         <CommonButton
           onClick={() => setMcqBankId(data._id)}
